backend/requirements.txt
# backend/requirements.txt
# NextStore E-commerce Backend Dependencies

# Django Core
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database - PostgreSQL
psycopg2-binary>=2.9.9

# CORS
django-cors-headers>=4.3.0

# Image Processing
Pillow>=10.0.0